requirements.txt
chainlit>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0

# LandingAI Document Parsing
agentic-doc>=0.1.0

# Vector Store
chromadb>=0.4.0

# OpenAI Embeddings
openai>=1.0.0
# LangChain Agent
langchain==0.2.14
langchain-openai==0.1.20
langchain-community==0.2.12
